Add ability to create prompts with a template format (#1193)
Ensure the template format on the prompt is used unless a user
explicitly overrides the rendering behavior. Regenerated types for the
sdk.

A user should be able to create a prompt with a template format:
```
const prompt = new Prompt(
      {
        id: "1",
        _xact_id: "xact_123",
        created: "2023-10-01T00:00:00Z",
        project_id: "project_123",
        prompt_session_id: "session_123",
        name: "test",
        slug: "test",
        prompt_data: {
          options: {
            model: "gpt-4o",
          },
          prompt: {
            type: "completion",
            content: "Complete this: {% if text %}{{text}}{% endif %}",
          },
          templateFormat: "nunjucks",
        },
      },
      {},
      true,
    );
```

A user should then have the template format respected during the build
rendering.