Make zod a peer dependency in the sdk (#1212)
Customers would like the ability to use the braintrust sdk with zod 4.
At the moment this is not possible due to a hard dependency on zod v3.

This PR starts to decouple zod version from the sdk by making it a peer
dependency and start the upgrade to zod v4.

Breaking Change
- zod will now need to be installed separately 

Typing issues:
- All types need to be kept on zod 3 in the sdk as all generated_types
from braintrust are only v3.